[Misc] Fix Some static_cast

+++ Xenon/Core/XCPU/Interpreter/PPC_ALU.cpp
@@ -24,7 +24,7 @@ struct addResult {
   // and out of bit[32] in 32 bit mode(of 64 bit implementations)
   addResult(T a, T b, bool sfBitMode) :
     result(a + b), carry(sfBitMode ? (result < a) :
-      (static_cast<u32>(result < a))) {}
+      (static_cast<u32>(result) < static_cast<u32>(a))) {}
 
   // Straightforward ADC with flags
   addResult(T a, T b, bool c, bool sfBitMode) :
